rename: book_storage 패키지명을 bookstorage로 변경
